ARM: dts: at91: add dts file for sama5d36ek CMP board
The difference between sama5d36ek_cmp.dts and sama5d36ek.dts
is listed as below.
 1. The USB host nodes are removed, that is, the USB host is disabled.
 2. The gpio_keys node is added to wake up from the sleep.
 3. The sama5d3xdm.dtsi isn't included, because the pins for LCD are
    conflicted with one for gpio_keys.
 4. Add pinctrl sleep state for adc0 node to fix the over consumption
    on VDDANA.

As said in errata, "When the USB host ports are used in high speed
mode (EHCI), it is not possible to suspend the ports if no device is
attached on each port. This leads to increased power consumption even
if the system is in a low power mode." That is why the the USB host
isn't enabled.

Signed-off-by: Wenyou Yang <wenyou.yang@atmel.com>
[nicolas.ferre@atmel.com: update with patch submitted to mainline]
Signed-off-by: Nicolas Ferre <nicolas.ferre@atmel.com>