mmc: sdhci: remove tuning failure log for DDR50 mode
In SD 3.0 spec, there is no tuning for the DDR50 mode. In SD 3.0.1 spec,
tuning can be performed for the DDR50 mode. Most of the SD card are
only compliant with SD 3.0 spec. It leads to get some logs because the
tuning procedure, which is not supported, has failed. This log is
useless most of the time.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>