net: macb: fix for suspend-to-mem feature
When suspending the macb, the management port for the MIIM interface was
disabled. When resuming, the PHY (KSZ8081RNB) was no longer responding.
We have to keep the management port enable to fix this issue.

Moreover, to get an up-to-date backup of the NCFGR register, it has to
be done just before suspending the macb.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
[claudiu.beznea@microchip.com: conform with latest driver version]
Signed-off-by: Claudiu Beznea <claudiu.beznea@microchip.com>