mmc: sdhci-of-at91: assigning clock frequency with DT
The IP is not integrated in the same way among all the Microchip
devices. The baseclock can be provided by the mainck or by the mulclk
divided by 2. Moreover, there is no way to know which frequency to
request for the mulclk as the content of the capabilities registers for
the base clk and mul are arbitrary values which don't reflect the
reality. To deal with this issue, the frequency of the mulclk is set in
the device tree with the help of assigned-clocks.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>