net: phy: micrel: make sure the factory test bit is cleared
The KSZ8081 PHY has a factory test mode which is set at the de-assertion
of the hard reset based on the TXC pin. If a pull-down is missing, or if
the pin has a pull-up, the factory test mode should be cleared by
manually writing a 0 (according to the datasheet). This patch makes sure
this factory test bit is cleared in config_init().

Tested-by: Alexandre Belloni <alexandre.belloni@bootlin.com>
Signed-off-by: Antoine Tenart <antoine.tenart@bootlin.com>
[nicolas.ferre@microchip.com: backport to 4.19]
Signed-off-by: Nicolas Ferre <nicolas.ferre@microchip.com>