ARM: dts: at91: sama5d27_wlsom1_ek: add serial line on mikro BUS 2
Add a new serial line for the mikro BUS 2 connector. It is connected to
ttyS2 in Linux and uses Flexcom0 as USART (RX/TX only).
Note that RTS/CTS pins of this USART could be found on the PIO USER
connector on the board: it can be handy to add them if needed.

Signed-off-by: Nicolas Ferre <nicolas.ferre@microchip.com>