drm/atmel-hlcdc: fix pixel clock divider selection
Comparison w/ a 10 times discriminator might be too much. We may end up
choosing a divider which goes to a result frequency to high departed
from the mode_rate. Choose the divisor which result in a frequency
closest to the mode_rate (either it is higher or lower than the mode_rate).

Fixes: f6f7ad323461 ("drm/atmel-hlcdc: allow selecting a higher pixel-clock than requested")