clk: at91: clk-generated: loop through all parents
clk_hw_get_num_parents() returns the number of parents of a hw clock.
Substracting with 1 and using "<" operator in for condition will
skip the last available parent for check. In case of GCK clocks with
changeable parents these would be used in case a previous best parent
one wasn't available.

Fixes: 1a1a36d72e3d3 ("clk: at91: clk-generated: make gclk determine audio_pll rate")
Signed-off-by: Claudiu Beznea <claudiu.beznea@microchip.com>