dmaegine: at_xdmac: add support for sama7g5 based at_xdmac
SAMA7G5 SoC uses a slightly different variant of the AT_XDMAC.
To avoid adding extra delay cycles by having a platform data struct
with extra offsets that need to be dynamically added in the
possible hot path of the DMA operations, created a new Kconfig
such that the SAMA7G5 specific offsets are actually inside
the defines, and the hot path is not affected, since all computations
regarding register offsets is done at compile time.
Added support by a new compatible and the defines are adapted
to the specific version considering the Kconfig option.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>