net: macb: add capability to not set the clock rate
SAMA7G5's ethernet IP has the possibility to be clocked by its
generic clock or by the external clock provided by the PHY.
The internal IP logic divides properly this clock depending on
the link speed. The patch adds a new capability so that
macb_set_tx_clock() to not be called for this IP when generic
clock is used (the clock rate is set at the boot time via
device tree and the driver only enables it). This is used
to avoid the following warning: "unable to generate target frequency"
when link goes up.

Fixes: b463343ec588 ("net: macb: remove clock code for sama7g5")
Signed-off-by: Claudiu Beznea <claudiu.beznea@microchip.com>