ARM: dts: at91: sama7g5: use assigned-clock-parents for SDMMC
Use assigned-clock-parents to PMC SYSPLL for the GCLK of the SDMMC blocks.
This will ensure that the SYSPLL is used as a parent clock source for the
GCLK. The GCLK must not be switched to another source, as this may lead
to glitches in the SDMMC block.
The bootloaders use SYSPLL so we must ensure that we do our best to use
SYSPLL for these clocks.

Suggested-by: Claudiu Beznea <claudiu.beznea@microchip.com>
Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>
Reviewed-by: Claudiu Beznea <claudiu.beznea@microchip.com>