spi: atmel-quadspi: Disable the QSPI IP at suspend()
It is safer to disable the QSPI IP at suspend, in order to avoid
possible impact of glitches on the internal FSMs. This is a theoretical
fix, there were no problems seen as of now. Tested on sama5d2 and
sam9x60 versions of the IP.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>
Link: https://lore.kernel.org/r/20210210135428.204134-1-tudor.ambarus@microchip.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit df6978b7ea6349eb32078c42b917559f5510aebd)
Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>