ARM: dts: at91: sama7g5ek: add pinctrl for ADC
By default the mikrobus AN pins have bias pull-down. This will interfere
with ADC measurements.
On a second thought, it would also be good to claim these pins for the ADC,
in pinctrl, even if the ADC IP itself does not require them to have a special
function.
This would avoid other blocks claim them, as these two pins are routed to the
Mikrobus connectors on this board.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>