Revert "mtd: spi-nor: core: Fix erase type discovery for overlaid region"
This reverts commit 6d6c7e7e9258e159e6977ff4694eb8efa226f2c3.

Instead of cherry-picking and solving conflicts for various commits
from SPI NOR that are needed to add support for the global unlock on
sst26 flashes, revert all the commits that come through fixes, merge
spi-nor/for-5.11 tag, and cherry-pick all the spi-nor commits up to
v5.13-rc1. All the fixes will be in-place, with a clear history. The
global unlock support for the sst26 flashes together with the octal dtr
support will come from mainline, without mangling the mainline patches.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>