crypto: atmel-aes: Add blocksize constraint for ECB and CBC modes
NIST 800-38A requires for the ECB and CBC modes that the total number
of bits in the plaintext to be a multiple of the block cipher.

-------
For CFB mode the requirement is that the plaintext to be a mutiple of
the data segments length. CFB's data segment legth is currently
represented as blocksize. Add these constraints for the ECB, CBC and
CFB modes.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>